Updated PackageItem Controller and all related classes and html pages.

+++ src/main/java/lv/pakit/repo/IPackageItemRepo.java
@@ -3,6 +3,6 @@ package lv.pakit.repo;
 import lv.pakit.model.PackageItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPackageItemRepo extends JpaRepository<PackageItem, Integer> {
+public interface IPackageItemRepo extends JpaRepository<PackageItem, Long> {
 
 }
